requirements.txt
streamlit
replicate
python-dotenv
requests
pillow
